refactor: move selection xml serializer to the extension method
