+++ src/Swashbuckle.AspNetCore.Filters/Examples/ResponseExample.cs
@@ -1,18 +1,10 @@
-using System;
-using System.IO;
 using System.Linq;
 using System.Collections.Generic;
 using Newtonsoft.Json.Serialization;
 using Newtonsoft.Json;
 using Microsoft.OpenApi.Models;
 using Microsoft.OpenApi.Any;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Formatters;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc.Infrastructure;
-using Microsoft.Net.Http.Headers;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Options;
 using Swashbuckle.AspNetCore.Filters.Extensions;
 
 namespace Swashbuckle.AspNetCore.Filters
@@ -53,14 +45,7 @@ namespace Swashbuckle.AspNetCore.Filters
                 OpenApiString xmlExample = null;
                 if (response.Value.Content.Keys.Any(k => k.Contains("xml")))
                 {
-                    if (outputFormatterSelector == null)
-                    {
-                        xmlExample = new OpenApiString(example.XmlSerialize());
-                    }
-                    else
-                    {
-                        xmlExample = WriteXml(outputFormatterSelector, example);
-                    }
+                    xmlExample = new OpenApiString(example.XmlSerialize(outputFormatterSelector));
                 }
 
                 foreach (var content in response.Value.Content)
@@ -75,52 +60,6 @@ namespace Swashbuckle.AspNetCore.Filters
                     }
                 }
             }
-        }
-
-        private OpenApiString WriteXml(OutputFormatterSelector outputFormatterSelector, object example)
-        {
-            using (var stringWriter = new StringWriter())
-            {
-                var outputFormatterContext = GetOutputFormatterContext(stringWriter, example, example.GetType());
-
-                var xmlFormatter = outputFormatterSelector.SelectFormatter(
-                    outputFormatterContext,
-                    new List<IOutputFormatter>(),
-                    new MediaTypeCollection());
-
-                xmlFormatter.WriteAsync(outputFormatterContext).GetAwaiter().GetResult();
-                stringWriter.FlushAsync().GetAwaiter().GetResult();
-                return new OpenApiString(stringWriter.ToString());
-            }
-        }
-
-        private OutputFormatterWriteContext GetOutputFormatterContext(
-            TextWriter writer,
-            object outputValue,
-            Type outputType,
-            string contentType = "application/xml; charset=utf-8")
-        {
-            return new OutputFormatterWriteContext(
-                GetHttpContext(contentType),
-                (stream, encoding) => writer,
-                outputType,
-                outputValue);
-        }
-
-        private static HttpContext GetHttpContext(string contentType)
-        {
-            var httpContext = new DefaultHttpContext();
-
-            httpContext.Request.Headers["Accept-Charset"] = MediaTypeHeaderValue.Parse(contentType).Charset.ToString();
-            httpContext.Request.ContentType = contentType;
-
-            httpContext.Response.Body = new MemoryStream();
-            httpContext.RequestServices =
-                new ServiceCollection()
-                    .AddSingleton(Options.Create(new MvcOptions()))
-                    .BuildServiceProvider();
-
-            return httpContext;
         }
     }
 }
